requirements.txt
python-telegram-bot 
langchain-openai 
langchain_core 
python-dotenv